Changed amount field - from Int to Float, 2 decimal places. Now it's possible to add float values.

+++ expense_tracker/models.py
@@ -5,7 +5,7 @@ from django.db import models
 # Model for saving up the expense
 class Expense(models.Model):
     name = models.CharField(max_length=100)
-    amount = models.PositiveIntegerField()
+    amount = models.DecimalField(max_digits=10, decimal_places=2)
     category = models.CharField(max_length=50)
     date = models.DateTimeField(auto_now=True)
